♻️ Refactor: api 호출로직 분리, 쿼리 캐시 개선
